Day 1 - Advent of Code
Completed in CoffeeScript.  Took a little while to think about the
problem in the (what I think is) the correct way.  Instead of thinking
of N/S/E/W, assigning them numbers and realizing that as you go around
the circle, the `num % 4` remains the same for each cardinal direction.

Will likely go back and clean things up and see what can be done better.
For now, happy with the working code.